Include names and sources even when empty
The names and sources fields are not optional according to the source maps spec. Omitting them when they are empty can cause source map parsers that expect them to always be present to crash. This changes the json encoding of those fields so they are no longer omitted when empty.

https://sourcemaps.info/spec.html